[feat/#18] :adhesive_bandage: fix: AmountCalculationUtil 유틸 클래스로 변환, 할인 금액, 수수료 계산 로직 수정
